requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic>=1.0.0
asyncpg>=0.27.0
attrs==25.1.0
discord.py>=2.0.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
psycopg2-binary>=2.9.0
pydantic>=1.0.0,<2.0.0
PyNaCl==1.5.0
python-dotenv>=0.19.0
SQLAlchemy>=2.0.0
typing_extensions==4.12.2
yarl==1.18.3
